fix a bug related to update more code dtandard remove Binary files in repo modified Method Length modified Package Name Style
