--- src/main/java/Duke/Tasks/TaskType.java
@@ -1,5 +1,0 @@
-package Duke.Tasks;
-
-public enum TaskType {
-    T, D, E, B
-}
