--- src/main/java/Duke/Utility/DukeException.java
@@ -1,7 +1,0 @@
-package Duke.Utility;
-
-public class DukeException extends Exception {
-    public DukeException(String Message) {
-        super(Message);
-    }
-}
