+++ src/main/java/duke/tasks/TaskType.java
@@ -1,0 +1,5 @@
+package duke.tasks;
+
+public enum TaskType {
+    T, D, E, B
+}
