+++ src/main/java/duke/utility/DukeException.java
@@ -1,0 +1,7 @@
+package duke.utility;
+
+public class DukeException extends Exception {
+    public DukeException(String Message) {
+        super(Message);
+    }
+}
