+++ src/test/java/duke/command/DeleteTaskTest.java
@@ -1,0 +1,11 @@
+package duke.command;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+
+/**
+ * Tests the functionality of the DeleteCommand class.
+ */
+public class DeleteCommandTest {
+
+}
